api/requirements.txt
flask
flask-cors
pandas
requests
python-dotenv
supabase
gunicorn 